ci: Test Rust code against current C code
Our current Rust CI was testing against the *stable* composefs,
i.e. whatever happened to be in the current Fedora.

This is too confusing, so move the CI for Rust into the main
workflow, and test against the C library under test too.

Also, add a `--all-features` run to add coverage for the
feature that enables running against the latest C library.

Signed-off-by: Colin Walters <walters@verbum.org>